feat(station): large WASM support for external canisters
