+++ core/station/api/src/external_canister.rs
@@ -3,6 +3,7 @@ use crate::{
     SortDirection, TimestampRfc3339, UuidDTO, ValidationMethodResourceTargetDTO,
 };
 use candid::{CandidType, Deserialize, Nat, Principal};
+use orbit_essentials::types::WasmModuleExtraChunks;
 
 pub type ExternalCanisterPermissionsInput = ExternalCanisterPermissionsDTO;
 
@@ -59,6 +60,7 @@ pub struct ChangeExternalCanisterOperationInput {
     pub mode: CanisterInstallMode,
     #[serde(with = "serde_bytes")]
     pub module: Vec<u8>,
+    pub module_extra_chunks: Option<WasmModuleExtraChunks>,
     #[serde(deserialize_with = "orbit_essentials::deserialize::deserialize_option_blob")]
     pub arg: Option<Vec<u8>>,
 }
