+++ core/station/impl/src/factories/requests/change_external_canister.rs
@@ -80,7 +80,7 @@ impl Execute for ChangeExternalCanisterRequestExecute<'_, '_> {
                 self.operation.input.canister_id,
                 self.operation.input.mode.clone(),
                 &self.operation.input.module,
-                &None,
+                &self.operation.input.module_extra_chunks,
                 self.operation.input.arg.clone(),
             )
             .await
