+++ core/station/impl/src/mappers/request_operation.rs
@@ -482,6 +482,7 @@ impl From<ChangeExternalCanisterOperationInput>
             canister_id: input.canister_id,
             mode: input.mode.into(),
             module: input.module,
+            module_extra_chunks: input.module_extra_chunks.map(|c| c.into()),
             arg: input.arg,
         }
     }
@@ -497,6 +498,7 @@ impl From<station_api::ChangeExternalCanisterOperationInput>
             canister_id: input.canister_id,
             mode: input.mode.into(),
             module: input.module,
+            module_extra_chunks: input.module_extra_chunks.map(|c| c.into()),
             arg: input.arg,
         }
     }
