+++ tools/dfx-orbit/src/args/request/canister.rs
@@ -179,6 +179,7 @@ impl RequestCanisterInstallArgs {
             canister_id,
             mode,
             module,
+            module_extra_chunks: None,
             arg,
         };
         Ok(RequestOperationInput::ChangeExternalCanister(operation))
